some minor nits
and a reminder of questions/issues raised offline:

1. the description of `spend_pubkey` in the first paycode table claims
"it is used to derive common secret" but `R = spend_pubkey` is not used
in the computation of the common secret `c`, according to the derivation
that follows. I only see `R` being used to derive `R'`, the pay to address.
Hence maybe the description of `spend_pubkey` needs slight adjustment.

2. the speak variously speaks of Bitcoin Cash and Bitcoin. This should
probably be unified and a note added that the spec is intended as a
proposal for both (if that is the case).